bp: Clear out bp utils
