+++ layers/best_practices/bp_image.cpp
@@ -133,3 +133,162 @@ bool BestPractices::PreCallValidateCreateImage(VkDevice device, const VkImageCre
 
     return skip;
 }
+
+void BestPractices::QueueValidateImageView(QueueCallbacks& funcs, Func command, vvl::ImageView* view,
+                                           IMAGE_SUBRESOURCE_USAGE_BP usage) {
+    if (view) {
+        auto image_state = std::static_pointer_cast<bp_state::Image>(view->image_state);
+        QueueValidateImage(funcs, command, image_state, usage, view->normalized_subresource_range);
+    }
+}
+
+void BestPractices::QueueValidateImage(QueueCallbacks& funcs, Func command, std::shared_ptr<bp_state::Image>& state,
+                                       IMAGE_SUBRESOURCE_USAGE_BP usage, const VkImageSubresourceRange& subresource_range) {
+    // If we're viewing a 3D slice, ignore base array layer.
+    // The entire 3D subresource is accessed as one atomic unit.
+    const uint32_t base_array_layer = state->createInfo.imageType == VK_IMAGE_TYPE_3D ? 0 : subresource_range.baseArrayLayer;
+
+    const uint32_t max_layers = state->createInfo.arrayLayers - base_array_layer;
+    const uint32_t array_layers = std::min(subresource_range.layerCount, max_layers);
+    const uint32_t max_levels = state->createInfo.mipLevels - subresource_range.baseMipLevel;
+    const uint32_t mip_levels = std::min(state->createInfo.mipLevels, max_levels);
+
+    for (uint32_t layer = 0; layer < array_layers; layer++) {
+        for (uint32_t level = 0; level < mip_levels; level++) {
+            QueueValidateImage(funcs, command, state, usage, layer + base_array_layer, level + subresource_range.baseMipLevel);
+        }
+    }
+}
+
+void BestPractices::QueueValidateImage(QueueCallbacks& funcs, Func command, std::shared_ptr<bp_state::Image>& state,
+                                       IMAGE_SUBRESOURCE_USAGE_BP usage, const VkImageSubresourceLayers& subresource_layers) {
+    const uint32_t max_layers = state->createInfo.arrayLayers - subresource_layers.baseArrayLayer;
+    const uint32_t array_layers = std::min(subresource_layers.layerCount, max_layers);
+
+    for (uint32_t layer = 0; layer < array_layers; layer++) {
+        QueueValidateImage(funcs, command, state, usage, layer + subresource_layers.baseArrayLayer, subresource_layers.mipLevel);
+    }
+}
+
+void BestPractices::QueueValidateImage(QueueCallbacks& funcs, Func command, std::shared_ptr<bp_state::Image>& state,
+                                       IMAGE_SUBRESOURCE_USAGE_BP usage, uint32_t array_layer, uint32_t mip_level) {
+    funcs.push_back([this, command, state, usage, array_layer, mip_level](const ValidationStateTracker& vst, const vvl::Queue& qs,
+                                                                          const vvl::CommandBuffer& cbs) -> bool {
+        ValidateImageInQueue(qs, cbs, command, *state, usage, array_layer, mip_level);
+        return false;
+    });
+}
+
+void BestPractices::ValidateImageInQueueArmImg(Func command, const bp_state::Image& image, IMAGE_SUBRESOURCE_USAGE_BP last_usage,
+                                               IMAGE_SUBRESOURCE_USAGE_BP usage, uint32_t array_layer, uint32_t mip_level) {
+    // Swapchain images are implicitly read so clear after store is expected.
+    const Location loc(command);
+    if (usage == IMAGE_SUBRESOURCE_USAGE_BP::RENDER_PASS_CLEARED && last_usage == IMAGE_SUBRESOURCE_USAGE_BP::RENDER_PASS_STORED &&
+        !image.IsSwapchainImage()) {
+        LogPerformanceWarning(
+            kVUID_BestPractices_RenderPass_RedundantStore, device, loc,
+            "%s %s Subresource (arrayLayer: %u, mipLevel: %u) of image was cleared as part of LOAD_OP_CLEAR, but last time "
+            "image was used, it was written to with STORE_OP_STORE. "
+            "Storing to the image is probably redundant in this case, and wastes bandwidth on tile-based "
+            "architectures.",
+            VendorSpecificTag(kBPVendorArm), VendorSpecificTag(kBPVendorIMG), array_layer, mip_level);
+    } else if (usage == IMAGE_SUBRESOURCE_USAGE_BP::RENDER_PASS_CLEARED && last_usage == IMAGE_SUBRESOURCE_USAGE_BP::CLEARED) {
+        LogPerformanceWarning(
+            kVUID_BestPractices_RenderPass_RedundantClear, device, loc,
+            "%s %s Subresource (arrayLayer: %u, mipLevel: %u) of image was cleared as part of LOAD_OP_CLEAR, but last time "
+            "image was used, it was written to with vkCmdClear*Image(). "
+            "Clearing the image with vkCmdClear*Image() is probably redundant in this case, and wastes bandwidth on "
+            "tile-based architectures.",
+            VendorSpecificTag(kBPVendorArm), VendorSpecificTag(kBPVendorIMG), array_layer, mip_level);
+    } else if (usage == IMAGE_SUBRESOURCE_USAGE_BP::RENDER_PASS_READ_TO_TILE &&
+               (last_usage == IMAGE_SUBRESOURCE_USAGE_BP::BLIT_WRITE || last_usage == IMAGE_SUBRESOURCE_USAGE_BP::CLEARED ||
+                last_usage == IMAGE_SUBRESOURCE_USAGE_BP::COPY_WRITE || last_usage == IMAGE_SUBRESOURCE_USAGE_BP::RESOLVE_WRITE)) {
+        const char* last_cmd = nullptr;
+        const char* vuid = nullptr;
+        const char* suggestion = nullptr;
+
+        switch (last_usage) {
+            case IMAGE_SUBRESOURCE_USAGE_BP::BLIT_WRITE:
+                vuid = kVUID_BestPractices_RenderPass_BlitImage_LoadOpLoad;
+                last_cmd = "vkCmdBlitImage";
+                suggestion =
+                    "The blit is probably redundant in this case, and wastes bandwidth on tile-based architectures. "
+                    "Rather than blitting, just render the source image in a fragment shader in this render pass, "
+                    "which avoids the memory roundtrip.";
+                break;
+            case IMAGE_SUBRESOURCE_USAGE_BP::CLEARED:
+                vuid = kVUID_BestPractices_RenderPass_InefficientClear;
+                last_cmd = "vkCmdClear*Image";
+                suggestion =
+                    "Clearing the image with vkCmdClear*Image() is probably redundant in this case, and wastes bandwidth on "
+                    "tile-based architectures. "
+                    "Use LOAD_OP_CLEAR instead to clear the image for free.";
+                break;
+            case IMAGE_SUBRESOURCE_USAGE_BP::COPY_WRITE:
+                vuid = kVUID_BestPractices_RenderPass_CopyImage_LoadOpLoad;
+                last_cmd = "vkCmdCopy*Image";
+                suggestion =
+                    "The copy is probably redundant in this case, and wastes bandwidth on tile-based architectures. "
+                    "Rather than copying, just render the source image in a fragment shader in this render pass, "
+                    "which avoids the memory roundtrip.";
+                break;
+            case IMAGE_SUBRESOURCE_USAGE_BP::RESOLVE_WRITE:
+                vuid = kVUID_BestPractices_RenderPass_ResolveImage_LoadOpLoad;
+                last_cmd = "vkCmdResolveImage";
+                suggestion =
+                    "The resolve is probably redundant in this case, and wastes a lot of bandwidth on tile-based architectures. "
+                    "Rather than resolving, and then loading, try to keep rendering in the same render pass, "
+                    "which avoids the memory roundtrip.";
+                break;
+            default:
+                break;
+        }
+
+        LogPerformanceWarning(
+            vuid, device, loc,
+            "%s %s Subresource (arrayLayer: %u, mipLevel: %u) of image was loaded to tile as part of LOAD_OP_LOAD, but last "
+            "time image was used, it was written to with %s. %s",
+            VendorSpecificTag(kBPVendorArm), VendorSpecificTag(kBPVendorIMG), array_layer, mip_level, last_cmd, suggestion);
+    }
+}
+
+void BestPractices::ValidateImageInQueue(const vvl::Queue& qs, const vvl::CommandBuffer& cbs, Func command, bp_state::Image& state,
+                                         IMAGE_SUBRESOURCE_USAGE_BP usage, uint32_t array_layer, uint32_t mip_level) {
+    auto queue_family = qs.queueFamilyIndex;
+    auto last_usage = state.UpdateUsage(array_layer, mip_level, usage, queue_family);
+
+    // Concurrent sharing usage of image with exclusive sharing mode
+    if (state.createInfo.sharingMode == VK_SHARING_MODE_EXCLUSIVE && last_usage.queue_family_index != queue_family) {
+        // if UNDEFINED then first use/acquisition of subresource
+        if (last_usage.type != IMAGE_SUBRESOURCE_USAGE_BP::UNDEFINED) {
+            // If usage might read from the subresource, as contents are undefined
+            // so write only is fine
+            if (usage == IMAGE_SUBRESOURCE_USAGE_BP::RENDER_PASS_READ_TO_TILE || usage == IMAGE_SUBRESOURCE_USAGE_BP::BLIT_READ ||
+                usage == IMAGE_SUBRESOURCE_USAGE_BP::COPY_READ || usage == IMAGE_SUBRESOURCE_USAGE_BP::DESCRIPTOR_ACCESS ||
+                usage == IMAGE_SUBRESOURCE_USAGE_BP::RESOLVE_READ) {
+                Location loc(command);
+                LogWarning(
+                    kVUID_BestPractices_ConcurrentUsageOfExclusiveImage, state.image(), loc,
+                    "Subresource (arrayLayer: %" PRIu32 ", mipLevel: %" PRIu32 ") of image is used on queue family index %" PRIu32
+                    " after being used on "
+                    "queue family index %" PRIu32
+                    ", "
+                    "but has VK_SHARING_MODE_EXCLUSIVE, and has not been acquired and released with a ownership transfer operation",
+                    array_layer, mip_level, queue_family, last_usage.queue_family_index);
+            }
+        }
+    }
+
+    // When image was discarded with StoreOpDontCare but is now being read with LoadOpLoad
+    if (last_usage.type == IMAGE_SUBRESOURCE_USAGE_BP::RENDER_PASS_DISCARDED &&
+        usage == IMAGE_SUBRESOURCE_USAGE_BP::RENDER_PASS_READ_TO_TILE) {
+        Location loc(command);
+        LogWarning(kVUID_BestPractices_StoreOpDontCareThenLoadOpLoad, device, loc,
+                   "Trying to load an attachment with LOAD_OP_LOAD that was previously stored with STORE_OP_DONT_CARE. This may "
+                   "result in undefined behaviour.");
+    }
+
+    if (VendorCheckEnabled(kBPVendorArm) || VendorCheckEnabled(kBPVendorIMG)) {
+        ValidateImageInQueueArmImg(command, state, last_usage.type, usage, array_layer, mip_level);
+    }
+}
